DSA_DPO/requirements.txt
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=2.3.1=pypi_0
accelerate=1.12.0=pypi_0
aiodns=3.5.0=py313h06a4308_1
aiohappyeyeballs=2.6.1=py313h06a4308_0
aiohttp=3.13.3=py313h47b2149_0
aiosignal=1.4.0=py313h06a4308_0
annotated-types=0.6.0=py313h06a4308_1
anyio=4.10.0=py313h06a4308_0
astunparse=1.6.3=pypi_0
attrs=25.4.0=py313h06a4308_2
bert-score=0.3.13=pypi_0
bitsandbytes=0.49.1=pypi_0
blis=1.3.3=pypi_0
bottleneck=1.6.0=pypi_0
brotlicffi=1.2.0.0=py313h7354ed3_0
bzip2=1.0.8=h5eee18b_6
c-ares=1.34.6=hd44998d_0
ca-certificates=2025.12.2=h06a4308_0
cachetools=5.5.1=py313h06a4308_0
catalogue=2.0.10=pypi_0
certifi=2026.01.04=py313h06a4308_0
cffi=2.0.0=py313h4eded50_1
charset-normalizer=3.4.4=py313h06a4308_0
clean-text=0.6.0=pypi_0
click=8.3.1=pypi_0
cloudpathlib=0.23.0=pypi_0
confection=0.1.5=pypi_0
contourpy=1.3.3=pypi_0
convokit=4.0.0=pypi_0
cryptography=46.0.3=py313h0a354b3_0
cuda-bindings=12.9.4=pypi_0
cuda-pathfinder=1.3.3=pypi_0
cut-cross-entropy=25.1.1=pypi_0
cycler=0.12.1=pypi_0
cymem=2.0.13=pypi_0
datasets=4.5.0=pypi_0
diffusers=0.36.0=pypi_0
dill=0.4.0=pypi_0
distro=1.9.0=pypi_0
dnspython=2.8.0=pypi_0
docstring-parser=0.17.0=pypi_0
emoji=1.7.0=pypi_0
en-core-web-sm=3.8.0=pypi_0
evaluate=0.4.6=pypi_0
filelock=3.20.3=pypi_0
flatbuffers=25.12.19=pypi_0
fonttools=4.61.1=pypi_0
frozenlist=1.8.0=py313hbdd6827_0
fsspec=2025.10.0=pypi_0
ftfy=6.3.1=pypi_0
gast=0.7.0=pypi_0
google-auth=2.47.0=py313h06a4308_0
google-genai=1.41.0=py313h06a4308_0
google-pasta=0.2.0=pypi_0
grpcio=1.76.0=pypi_0
h11=0.16.0=py313h06a4308_1
h5py=3.12.1=pypi_0
hf-transfer=0.1.9=pypi_0
hf-xet=1.2.0=pypi_0
httpcore=1.0.9=py313h06a4308_0
httpx=0.28.1=py313h06a4308_1
huggingface-hub=0.36.0=pypi_0
idna=3.11=py313h06a4308_0
importlib-metadata=8.7.1=pypi_0
jinja2=3.1.6=pypi_0
jiter=0.12.0=pypi_0
joblib=1.5.3=pypi_0
keras=3.13.1=pypi_0
kiwisolver=1.4.9=pypi_0
ld_impl_linux-64=2.44=h153f514_2
libclang=18.1.1=pypi_0
libexpat=2.7.3=h7354ed3_4
libffi=3.4.4=h6a678d5_1
libgcc=15.2.0=h69a1729_7
libgcc-ng=15.2.0=h166f726_7
libgomp=15.2.0=h4751f2c_7
libmpdec=4.0.0=h5eee18b_0
libstdcxx=15.2.0=h39759b7_7
libstdcxx-ng=15.2.0=hc03a8fd_7
libuuid=1.41.5=h5eee18b_0
libuv=1.48.0=h5eee18b_0
libxcb=1.17.0=h9b100fa_0
libzlib=1.3.1=hb25bd0a_0
markdown=3.10.1=pypi_0
markdown-it-py=4.0.0=pypi_0
markupsafe=3.0.3=pypi_0
matplotlib=3.10.8=pypi_0
mdurl=0.1.2=pypi_0
ml-dtypes=0.5.4=pypi_0
mpmath=1.3.0=pypi_0
msgpack=1.1.2=pypi_0
msgpack-numpy=0.4.8=pypi_0
msgspec=0.20.0=pypi_0
multidict=6.7.0=py313hee96239_0
multiprocess=0.70.18=pypi_0
murmurhash=1.0.15=pypi_0
namex=0.1.0=pypi_0
ncurses=6.5=h7934f7d_0
networkx=3.6.1=pypi_0
nltk=3.9.2=pypi_0
numexpr=2.14.1=pypi_0
numpy=2.4.1=pypi_0
nvidia-cublas-cu12=12.8.4.1=pypi_0
nvidia-cuda-cupti-cu12=12.8.90=pypi_0
nvidia-cuda-nvrtc-cu12=12.8.93=pypi_0
nvidia-cuda-runtime-cu12=12.8.90=pypi_0
nvidia-cudnn-cu12=9.10.2.21=pypi_0
nvidia-cufft-cu12=11.3.3.83=pypi_0
nvidia-cufile-cu12=1.13.1.3=pypi_0
nvidia-curand-cu12=10.3.9.90=pypi_0
nvidia-cusolver-cu12=11.7.3.90=pypi_0
nvidia-cusparse-cu12=12.5.8.93=pypi_0
nvidia-cusparselt-cu12=0.7.1=pypi_0
nvidia-nccl-cu12=2.27.5=pypi_0
nvidia-nvjitlink-cu12=12.8.93=pypi_0
nvidia-nvshmem-cu12=3.4.5=pypi_0
nvidia-nvtx-cu12=12.8.90=pypi_0
openai=2.15.0=pypi_0
openssl=3.0.18=hd6dcaed_0
opt-einsum=3.4.0=pypi_0
optree=0.18.0=pypi_0
packaging=26.0=pypi_0
pandas=3.0.0=pypi_0
peft=0.18.1=pypi_0
pillow=12.1.0=pypi_0
pip=25.3=pyhc872135_0
preshed=3.0.12=pypi_0
propcache=0.4.1=py313h47b2149_0
protobuf=6.33.4=pypi_0
psutil=7.2.1=pypi_0
pthread-stubs=0.3=h0ce48e5_1
pyarrow=23.0.0=pypi_0
pyasn1=0.6.1=py313h06a4308_0
pyasn1-modules=0.4.2=py313h06a4308_0
pycares=4.10.0=py313ha137689_0
pycparser=2.23=py313h06a4308_0
pydantic=2.12.4=py313h06a4308_0
pydantic-core=2.41.5=py313h498d7c9_1
pygments=2.19.2=pypi_0
pymongo=4.16.0=pypi_0
pyopenssl=25.3.0=py313hef88997_0
pyparsing=3.3.2=pypi_0
pysocks=1.7.1=py313h06a4308_1
python=3.13.11=hcf712cf_100_cp313
python-dateutil=2.9.0.post0=pypi_0
python_abi=3.13=3_cp313
pyyaml=6.0.3=pypi_0
readline=8.3=hc2a1206_0
regex=2026.1.15=pypi_0
requests=2.32.5=py313h06a4308_1
rich=14.2.0=pypi_0
rouge-score=0.1.2=pypi_0
rsa=4.7.2=pyhd3eb1b0_1
ruff=0.14.14=pypi_0
safetensors=0.7.0=pypi_0
scikit-learn=1.8.0=pypi_0
scipy=1.17.0=pypi_0
sentence-transformers=5.2.0=pypi_0
sentencepiece=0.2.1=pypi_0
setuptools=80.9.0=py313h06a4308_0
six=1.17.0=pypi_0
smart-open=7.5.0=pypi_0
sniffio=1.3.1=py313h06a4308_0
spacy=3.8.11=pypi_0
spacy-legacy=3.0.12=pypi_0
spacy-loggers=1.0.5=pypi_0
sqlite=3.51.1=he0a8d7e_0
srsly=2.5.2=pypi_0
sympy=1.14.0=pypi_0
tenacity=9.1.2=py313h06a4308_1
tensorboard=2.20.0=pypi_0
tensorboard-data-server=0.7.2=pypi_0
tensorflow=2.20.0=pypi_0
termcolor=3.3.0=pypi_0
tf-keras=2.20.1=pypi_0
thinc=8.3.10=pypi_0
threadpoolctl=3.6.0=pypi_0
tk=8.6.15=h54e0aa7_0
tokenizers=0.22.2=pypi_0
torch=2.10.0=pypi_0
torchao=0.15.0=pypi_0
torchvision=0.25.0=pypi_0
tqdm=4.67.1=pypi_0
transformers=4.57.2=pypi_0
triton=3.6.0=pypi_0
trl=0.23.0=pypi_0
typeguard=4.4.4=pypi_0
typer-slim=0.21.1=pypi_0
typing-extensions=4.15.0=py313h06a4308_0
typing-inspection=0.4.2=py313h06a4308_0
typing_extensions=4.15.0=py313h06a4308_0
tyro=1.0.5=pypi_0
tzdata=2025c=he532380_0
unidecode=1.4.0=pypi_0
unsloth=2025.11.1=pypi_0
unsloth-zoo=2025.11.2=pypi_0
urllib3=2.6.3=py313h06a4308_0
uvloop=0.22.1=py312h939ff40_1
wasabi=1.1.3=pypi_0
wcwidth=0.3.2=pypi_0
weasel=0.4.3=pypi_0
websockets=15.0.1=py313h5eee18b_0
werkzeug=3.1.5=pypi_0
wheel=0.45.1=py313h06a4308_0
wrapt=2.0.1=pypi_0
xformers=0.0.34=pypi_0
xorg-libx11=1.8.12=h9b100fa_1
xorg-libxau=1.0.12=h9b100fa_0
xorg-libxdmcp=1.1.5=h9b100fa_0
xorg-xorgproto=2024.1=h5eee18b_1
xxhash=3.6.0=pypi_0
xz=5.6.4=h5eee18b_1
yarl=1.22.0=py313hee96239_0
zipp=3.23.0=pypi_0
zlib=1.3.1=hb25bd0a_0
